requirements.txt
aerich==0.5.8
aiosqlite==0.17.0
anyio==3.3.3
asgiref==3.4.1
bidict==0.21.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
ddlparse==1.10.0
dictdiffer==0.9.0
fastapi==0.70.0
fastapi-pagination==0.9.1
h11==0.12.0
idna==3.3
importlib-metadata==4.8.2
iso8601==0.1.16
jmespath==0.10.0
Pillow==8.4.0
pycparser==2.21
pydantic==1.8.2
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==2.4.7
pypika-tortoise==0.1.1
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2021.3
qrcode==7.3.1
requests==2.26.0
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
tortoise-orm==0.17.8
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.16.0
zipp==3.6.0
aiofiles~=0.8.0